requirements.txt
supabase
pywin32
python-dotenv
